% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showstats.R
\name{showstats}
\alias{showstats}
\title{Show Statistics}
\usage{
showstats(type)
}
\arguments{
\item{type}{The kind of analysis to be done. The possible options are \code{graduates}, \code{latin honors}, \code{department honors}, \code{latin honors detail} and \code{department honors detail}. \code{graduates} displays the timeplot of the gender distribution of the graduating population. \code{latin honors} displays the timeplot of the gender distribution of the population that has latin honors. \code{latin honors detail} splits the population that has latin honors into three: Summa Cum Laude, Magna Cum Laude and Cum Laude, and displays the timeplot of the gender distribution. \code{department honors} displays the timeplot of the gender distribution of the population that has department honors. \code{department honors detail} splits the population that has department honors to two: honors and highest honors, and displays the timeplot of the gender distribution.}
}
\value{
Results from the kind of analysis specified by \code{type}.
}
\description{
Uses the data from the read function and performs appropriate analysis and generates graphics associated with the gender distribution.
}
